refactor(compute/sim): Use workload chaining for boot delay
This change updates the implementation of `SimHost` to use workload
chaining for modelling boot delays. Previously, this was implemented by
sleeping 1 millisecond using Kotlin coroutines. With this change, we
remove the need for coroutines and instead use the `SimDurationWorkload`
to model the boot delay.

In the future, we envision a user-supplied stochastic boot model to
model the boot delay for VM instances.