fix(web/ui): Disable configuration of basePath
This change removes the ability to configure the basePath of the Next.js
application using the Quarkus extension. This functionality was brittle
due to relying on Next.js internals coping with out replacement
strategy.

We should wait for Next.js to implement proper support for
changing the base path at runtime before making this functionality
available again.